fix: scope manager
